Add files via upload
MobilePhoneNumber + some fixes

+++ src/main/java/com/presentation/demo/controller/BillController.java
@@ -36,7 +36,7 @@ public class BillController {
 
     @GetMapping("/createbill/{id}")
     @ResponseBody
-    public String createBill(@PathVariable("id") Long id) {
+    public String createBill(@PathVariable("id") Integer id) {
 
         Random rand = new Random();
         Bill bill = new Bill();
