+++ src/main/java/com/presentation/demo/repository/AuthenticatedUserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface AuthenticatedUserRepository extends JpaRepository<User,Long> {
-    User findAuthenticatedUserById(Long id);
+    User findAuthenticatedUserById(Integer id);
     User findAuthenticatedUserByUsername(String username);
 }
