+++ src/main/java/com/presentation/demo/service/user/authorized/AuthenticatedUsersService.java
@@ -5,7 +5,7 @@ import com.presentation.demo.model.User;
 import java.util.List;
 
 public interface AuthenticatedUsersService {
-    User findAuthenticatedUserByID(Long id);
+    User findAuthenticatedUserByID(Integer id);
     User findAuthenticatedUserByUsername(String username);
     void save(User user);
     void delete(User user);
